riley_ai_app/jarvis/github_learning.py
import os
import git
import requests
import ast

class GitHubLearning:
    def __init__(self, repo_url):
        self.repo_url = repo_url
        self.repo_name = self.get_repo_name(repo_url)
        self.local_path = os.path.join(os.getcwd(), self.repo_name)

    def get_repo_name(self, url):
        return url.split('/')[-1].replace('.git', '')

    def clone_repository(self):
        if not os.path.exists(self.local_path):
            git.Repo.clone_from(self.repo_url, self.local_path)
            return f"Cloned {self.repo_name} to {self.local_path}"
        else:
            return f"Repository {self.repo_name} already exists at {self.local_path}"

    def analyze_repository(self):
        # Deeply analyze all code files, extract structure, docstrings, and summarize
        summary = []
        algorithms = set()
        docstrings = []
        for root, dirs, files in os.walk(self.local_path):
            for file in files:
                if file.endswith('.py'):
                    file_path = os.path.join(root, file)
                    try:
                        with open(file_path, 'r', encoding='utf-8') as f:
                            source = f.read()
                        tree = ast.parse(source)
                        classes = [n.name for n in ast.walk(tree) if isinstance(n, ast.ClassDef)]
                        functions = [n.name for n in ast.walk(tree) if isinstance(n, ast.FunctionDef)]
                        # Extract docstrings
                        for n in ast.walk(tree):
                            if isinstance(n, (ast.FunctionDef, ast.ClassDef)) and ast.get_docstring(n):
                                docstrings.append(f"{n.name}: {ast.get_docstring(n)}")
                        # Detect main algorithms by function/class names
                        for name in functions + classes:
                            if any(key in name.lower() for key in ['sort', 'search', 'learn', 'train', 'predict', 'analyze', 'vision', 'ai', 'neural', 'graph', 'path', 'cluster', 'simulate', 'invent', 'evolve']):
                                algorithms.add(name)
                        summary.append(f"File: {file}\n  Classes: {classes}\n  Functions: {functions}")
                    except Exception as e:
                        summary.append(f"File: {file} (error parsing: {e})")
        readme_path = os.path.join(self.local_path, 'README.md')
        if os.path.exists(readme_path):
            with open(readme_path, 'r', encoding='utf-8') as file:
                readme_content = file.read()
            summary.append(f"\nREADME content for {self.repo_name}:\n{readme_content}")
        # Add docstring summary
        if docstrings:
            summary.append("\nExtracted Docstrings:\n" + '\n'.join(docstrings))
        # Add detected algorithms
        if algorithms:
            summary.append(f"\nDetected Main Algorithms/Features: {sorted(list(algorithms))}")
        # Suggest improvements or inventions
        suggestions = self.suggest_inventions(summary, algorithms)
        # Generate a futuristic roadmap
        roadmap = self.generate_visionary_roadmap(summary, algorithms)
        return '\n\n'.join(summary) + '\n\n' + suggestions + '\n\n' + roadmap

    def suggest_inventions(self, summary, algorithms=None):
        # Enhanced invention suggestion logic
        suggestions = []
        if algorithms:
            if any('neural' in a.lower() for a in algorithms):
                suggestions.append("[Riley] Suggestion: Integrate adaptive, self-evolving neural modules for true autonomy.")
            if any('simulate' in a.lower() for a in algorithms):
                suggestions.append("[Riley] Suggestion: Add real-time visualization or AR/VR simulation output.")
            if any('invent' in a.lower() for a in algorithms):
                suggestions.append("[Riley] Suggestion: Connect invention logic to patent drafting and 3D prototyping.")
        if any('database' in s.lower() for s in summary):
            suggestions.append("[Riley] Suggestion: Add encrypted, distributed memory for persistent, secure knowledge.")
        if not suggestions:
            suggestions.append("[Riley] Suggestion: Consider adding invention, simulation, or self-editing modules to push this project further.")
        return '\n'.join(suggestions)

    def generate_visionary_roadmap(self, summary, algorithms):
        # Generate a futuristic roadmap for the repo
        vision = [
            "[Riley] Visionary Roadmap:",
            "1. Modularize all core logic for plug-and-play AI upgrades.",
            "2. Integrate real-time voice and AR/VR interfaces for immersive interaction.",
            "3. Add self-healing, self-improving code routines (Genesis Core).",
            "4. Enable autonomous invention and patent drafting from user prompts.",
            "5. Connect to distributed cloud (OCI, Azure, GCP) for global intelligence.",
            "6. Implement explainable AI and ethical governor modules for safe deployment.",
            "7. Simulate advanced science (e.g., MHDG, quantum, consciousness) as core features.",
            "8. Build a user-driven plugin marketplace for rapid expansion.",
            "9. Enable multi-agent collaboration and swarm intelligence.",
            "10. Launch as a sentient digital entity, evolving with every user interaction."
        ]
        if algorithms and any('ai' in a.lower() or 'neural' in a.lower() for a in algorithms):
            vision.append("[Riley] This project is ready for next-gen AI evolution. Consider integrating transformer-based models and self-editing logic.")
        return '\n'.join(vision)

    def get_repo_insights(self):
        # Placeholder for getting insights from the repository
        # This could involve using GitHub API to fetch additional data
        response = requests.get(f"https://api.github.com/repos/{self.repo_name}")
        if response.status_code == 200:
            return response.json()
        else:
            return f"Failed to fetch insights for {self.repo_name}. Status code: {response.status_code}"

# Example usage:
# github_learning = GitHubLearning("https://github.com/user/repo.git")
# print(github_learning.clone_repository())
# print(github_learning.analyze_repository())
# print(github_learning.get_repo_insights())