riley_ai_app/requirements.txt
pyqt6
openai
speechrecognition
pyttsx3
gitpython
requests
pysqlite3
PyYAML
tqdm